Fix showing next episode
Selecting next episode to watch was done using the aired date. This has the
drawback that if two episodes share the same aired date, the last one was
choosen. Also, those episodes with no aired date were discarded.

This patch fixes this issue by selecting the next episode using the
chronological order of episodes: the first unseen episode for season 1, then
for season 2, and so on. And at the end, the Special season.

This works even for those episodes that do not have an aired date.