Save on database on tmp file first
This is a way of reducing the change of having some problem when saving
database, specially if it is too big.

DB is saved on tmp file and only when it is saved completely on disk,
moved to the final destination. This movement should be greatly faster
than the save itself, so if something wrong happens during the save,
the old DB would be available, but at least it won't be corrupted.