Check if SeriesFinale is already running
When launching SF, it checks if a previous instance is still running. If so, it
shows a banner telling it and waits until the previous instance finalizes.

Users can close the waiting instance if they launched it by mistake.

This avoids the problem of having two running instances, that can potentially
overwrite the same DB and therefore loose user changes.