Fix header on Series page
In the Series page, it is shown "SeriesFinale" + update button, and the
list of available series.

When scrolling the series, it turns out that the header is also scrolled
too.

This commit makes fixes the header, so if user wants to update all the
series, they don't need to scroll to the first series to see the Update
button.